support dantai hokei/tenkai winners in api
